refactor: separated utils and handler
